+++ inst/modules/all_urls.R
@@ -26,7 +26,7 @@ all_urls <- function(paper, ...) {
   # return a list ----
   list(
     table = table,
-    summary = summary_table,
+    summary_table = summary_table,
     na_replace = 0,
     traffic_light = tl
   )