+++ inst/modules/exact_p.R
@@ -15,7 +15,7 @@
 #' @examples
 #' module_run(psychsci, "exact_p")
 exact_p <- function(paper, ...) {
-  # ---- Detailed table of results ----
+  # table ----
   p <- module_run(paper, "all_p_values")$table
 
   # Expand the sentences so the full sentence can be seen
@@ -30,73 +30,63 @@ exact_p <- function(paper, ...) {
   p$imprecise <- p$imprecise | !p$p_comp %in% c("=", "<")
   p$imprecise <- p$imprecise | is.na(p$p_value)
 
-  cols <- c("expanded")
-  table <- p[p$imprecise, cols]
+  # remove false positive "*p < .05"
+  star_pattern <- "\\*\\s*p\\s*<\\s*0?\\.0+[15]"
+  stars <- grepl(star_pattern, p$expanded)
+  p$imprecise <- p$imprecise & !stars
 
-  # summary output for paperlists ----
-  summary_table <- dplyr::filter(p, imprecise == TRUE)
+  cols <- c("text", "expanded")
+  report_table <- unique(p[p$imprecise, cols, drop = FALSE])
+  colnames(report_table) <- c("P-Value", "Sentence")
+
+  # summary_table ----
+  summary_table <- p[p$imprecise, , drop = FALSE]
   summary_table <- dplyr::count(summary_table, id, name = "n_imprecise")
 
-  # ---- Determine traffic light ----
-  if(nrow(table) == 0) {
+  # traffic light ----
+  if (nrow(p) == 0) {
     tl <- "na"
-    } else {
+  } else if(nrow(report_table) == 0) {
+    tl <- "green"
+  } else {
     tl <- "red"
-    }
+  }
 
-  # ---- Build report ----
-  if (nrow(table) == 0) {
-    report <- "We detected no imprecise *p* values."
+  # report / summary_text ----
+  if (tl == "na") {
+    report <- "We detected no *p* values."
+    summary_text <- report
+  } else if (tl == "green") {
+    report <- "All detected *p* values were reported with appropriate precision."
+    summary_text <- report
   } else {
     module_output <- sprintf(
-      "We found %d imprecise *p* values. Reporting *p* values imprecisely (e.g., *p* < .05) reduces transparency, reproducibility, and re-use (e.g., in *p* value meta-analyses). Best practice is to report exact p-values with three decimal places (e.g., *p* = .032) unless *p* values are smaller than 0.001, in which case you can use *p* < .001.",
-      nrow(table)
+      "We found %d imprecise *p* value%s. Reporting *p* values imprecisely (e.g., *p* < .05) reduces transparency, reproducibility, and re-use (e.g., in *p* value meta-analyses). Best practice is to report exact p-values with three decimal places (e.g., *p* = .032) unless *p* values are smaller than 0.001, in which case you can use *p* < .001.",
+      nrow(report_table), ifelse(nrow(report_table) == 1, "", "s")
     )
 
-    # Combine problematic sentences
-    issues_found <- paste(sprintf("%s", table$expanded), collapse = "\n\n")
+    summary_text <- sprintf("We found %d imprecise *p* value%s.",
+                            nrow(report_table),
+                            ifelse(nrow(report_table) == 1, "", "s"))
 
-    # Scrollable block for sentences
-    sentences_block <- paste0(
-      "<strong><span style='font-size:20px; color:#006400;'>Imprecise p values</span></strong>",
-      "<div style='border:1px solid #ccc; padding:10px; ",
-      "max-height:250px; overflow-y:auto; background-color:#f9f9f9; ",
-      "margin-top:5px; margin-bottom:15px;'>",
-      "<ul style='list-style-type: circle; padding-left:20px; margin:0;'>",
-      issues_found,
-      "</ul>",
-      "</div>"
-    )
-
-    # Guidance block
-    guidance <- paste0(
+    # Guidance text
+    guidance <- c(
       "The APA manual states: Report exact *p* values (e.g., *p* = .031) to two or three decimal places. However, report *p* values less than .001 as *p* < .001. However, 2 decimals is too imprecise for many use-cases (e.g., a *p* value meta-analysis), so report *p* values with three digits.",
       "American Psychological Association. (2020). Publication manual of the American Psychological Association 2020: the official guide to APA style (7th ed.). American Psychological Association."
     )
 
-    guidance_block <- paste0(
-      "<details style='display:inline-block;'>",
-      "<summary style='cursor:pointer; margin:0; padding:0;'>",
-      "<strong><span style='font-size:20px; color:#006400;'>Learn More</span></strong>",
-      "</summary>",
-      "<div style='margin-top:10px;'>",
-      guidance,
-      "</div>",
-      "</details>"
-    )
-
-    # Combine everything into report
-    report <- sprintf(
-      "%s\n\n%s\n\n%s",
-      module_output, sentences_block, guidance_block
-    )
+    # Combine everything into report text
+    report <- c(module_output,
+                scroll_table(report_table, colwidths = c(.1, .9)),
+                collapse_section(guidance))
   }
 
   # ---- Return list ----
   list(
-    traffic_light = tl,
-    report = report,
     table = p,
-    summary_table = summary_table
+    summary_table = summary_table,
+    traffic_light = tl,
+    summary_text = summary_text,
+    report = report
   )
 }